services/phone_agent/requirements.txt
# Meridian AI Phone Agent — 100% Open Source Dependencies
# No paid API SDKs required

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Voice pipeline
pipecat-ai>=0.0.40

# STT — WhisperLiveKit + faster-whisper (replaces Deepgram)
whisperlivekit>=0.3.0
faster-whisper>=1.0.0

# TTS — Kokoro (replaces Cartesia)
kokoro>=0.9.4

# LLM — Ollama client (replaces Anthropic/OpenAI SDK)
aiohttp>=3.9.0

# Audio processing
numpy>=1.26.0
scipy>=1.12.0

# VAD
silero-vad>=5.0

# HTTP client for POS APIs + Supabase
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
